# GSdx-HW: Optimize RT buffer size calculation for custom resolutions