@gregory38 2017-05-12
Anyway, I'm not sure it is a good solution (but I need to understand what you did first). Maybe it would be possible to dynamically allocate some buffers.

The constrain is that size of the rendering on the GPU is intersection of all buffers (color + depth).

We can monitor the real size of the GS draw call, Here a kind of algo (might need some tuning)
* if depth enabled
  * if cache hit, increase the size of the buffer is needed (copy)
  * else allocate both (color depth) buffers based on max FBW/Display BW values ever reported
* else
  * if cache hit, increase the size of the buffer is needed (copy)
  * else if sprite rendering and draw is small (let's say 128x128) allocate a RT of 128x128
  * else allocate a buffer based on max FBW/Display BW values ever reported