@lightningterror 2017-05-19
> I'm not sure if your idea will be reliable, there are too much assumptions on it which makes the algorithm seem a bit volatile, though I haven't really looked much into it (lack of free time) and plus not too keen on adding a new algorithm just before a release, don't want too much regressions on my head. ;)

There's gonna be plenty of other releases for that, save it for 1.7 xD