@gregory38 2017-05-25
On Mon, 22 May 2017 09:28:40 -0700
Akash <notifications@github.com> wrote:

> > But I guess most of your gain come from it. What happen if you keep 640 instead.  
> 
> It'll work but will consume more memory than necessary, like cases at ICO where scissor is 512 and the output circuit height is 256. And also similar situations at VP2/SH2 where output circuit is 448 but scissor 512 (needed for proper scaling).
> 
> I think using the scissor should be safe here, we have a safe lower limit (output circuit height) and a safe upper limit (640), nothing could possibly go wrong here. We're only using the scissor when it has a sensible value.
> 
> 

It will work most of the time because most games will use latest frame buffer as image source and will use a scissor at least on the framebuffer size.

But you don't have any guarantee.

However if I understand it right, the size is only increased. So we can suspect the value of scissor will be big enough at least once.