@gregory38 2017-05-27
@ssakash I let you handle the tutleli feedback and then we could merge it.