@ssakash 2017-06-01
> Maybe removes the use std statement and fixes all missing std:: in a future commit :)

I figure @turtleli would be more interested, I don't really care much about it.