@turtleli 2017-06-01
Yeah, I've already removed `using namespace std;` in one of my branches ;)