@lightningterror 2017-06-04
1-2 fps difference in GT4. More fps is always welcome.